fix supabase schema errors and adjust tests
