+++ package.json
@@ -70,9 +70,11 @@
     "eslint": "^9",
     "eslint-config-next": "15.3.3",
     "jest": "^29.7.0",
+    "jest-environment-jsdom": "^30.0.0-beta.3",
     "postcss": "^8.5.4",
     "tailwindcss": "^4.1.8",
     "ts-jest": "^29",
+    "ts-node": "^10.9.2",
     "tw-animate-css": "^1.3.3",
     "typescript": "^5"
   }
